#128: 재치등록_Second ViewController, View 파일 분리
